+++ Server/sel.sales/src/main/java/view/composites/EditDishListener.java
@@ -1,14 +1,16 @@
 package view.composites;
 
+import java.math.BigDecimal;
+
 import controller.BusinessEvent;
 import controller.IBusinessEventShooter;
 import controller.IController;
-import model.IDishMenuItemData;
-import model.IDishMenuItemDataFactory;
-import model.IDishMenuItemID;
-import model.IDishMenuItemIDFactory;
-import view.repository.ClickEventListener;
+import model.dish.IDishMenuItemData;
+import model.dish.IDishMenuItemDataFactory;
+import model.dish.IDishMenuItemID;
+import model.dish.IDishMenuItemIDFactory;
 import view.repository.HasText;
+import view.repository.uiwrapper.ClickEventListener;
 
 public class EditDishListener extends ClickEventListener implements IBusinessEventShooter {	private IController controller;
 	private HasText dishName;
@@ -35,9 +37,9 @@ public class EditDishListener extends ClickEventListener implements IBusinessEve
 		
 		IDishMenuItemData data = fac.constructData(
 				this.getDishName().getText(),
-				Double.valueOf(this.getPortion().getText()).doubleValue(), 
-				Double.valueOf(this.getPrice().getText()).doubleValue(),
-				Double.valueOf(this.getProductionCost().getText()).doubleValue(),
+				BigDecimal.valueOf(Double.valueOf(this.getPortion().getText()).doubleValue()), 
+				BigDecimal.valueOf(Double.valueOf(this.getPrice().getText()).doubleValue()),
+				BigDecimal.valueOf(Double.valueOf(this.getProductionCost().getText()).doubleValue()),
 				this.getDishID().getText(),
 				idFac
 		);
